refactor: counsel 관련 엔티티 분리
