+++ src/services/counselings/aggregates/counselors/infrastructures/adaptors/psql.counselors.repository.adaptor.ts
@@ -1,11 +1,8 @@
 import { UniqueEntityId } from "~shared/core/domain/UniqueEntityId";
-import { CounselorsEntity } from "~shared/core/infrastructure/entities/Counselors.entity";
+import { CounselorsEntity } from "~shared/core/infrastructure/entities/counsels/Counselors.entity";
 import { Counselors } from "~counselings/aggregates/counselors/domain/counselors";
 import { PsqlCounselorsMapper } from "~counselings/aggregates/counselors/infrastructures/adaptors/mapper/psql.counselors.mapper";
-import {
-  CounselorsRepositoryPort,
-  FindManyPropsInCounselorsRepository,
-} from "~counselings/aggregates/counselors/infrastructures/counselors.repository.port";
+import { CounselorsRepositoryPort, FindManyPropsInCounselorsRepository } from "~counselings/aggregates/counselors/infrastructures/counselors.repository.port";
 
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
@@ -42,9 +39,7 @@ export class PsqlCounselorsRepositoryAdaptor implements CounselorsRepositoryPort
 
   async findAll(): Promise<Counselors[]> {
     const counselorsEntities = await this.counselorsRepository.find();
-    return counselorsEntities
-      .map((counselorsEntity) => PsqlCounselorsMapper.toDomain(counselorsEntity))
-      .filter((counselor) => counselor !== null);
+    return counselorsEntities.map((counselorsEntity) => PsqlCounselorsMapper.toDomain(counselorsEntity)).filter((counselor) => counselor !== null);
   }
 
   async findMany(props: FindManyPropsInCounselorsRepository): Promise<Counselors[]> {
@@ -59,8 +54,6 @@ export class PsqlCounselorsRepositoryAdaptor implements CounselorsRepositoryPort
     const counselorsEntities: CounselorsEntity[] = await this.counselorsRepository.find({
       where: findOptionsWhere,
     });
-    return counselorsEntities
-      .map((counselorsEntity) => PsqlCounselorsMapper.toDomain(counselorsEntity))
-      .filter((counselor) => counselor !== null);
+    return counselorsEntities.map((counselorsEntity) => PsqlCounselorsMapper.toDomain(counselorsEntity)).filter((counselor) => counselor !== null);
   }
 }
